pyproject.toml
[tool.poetry]
name = "wit"
version = "0.1.0"
description = "Wikipedia Image/Caption Matching"
authors = ["SeanCondon <sean.condon@intel.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.0"
pillow = "^11.0.0"
torchvision = "^0.20.0"
transformers = "^4.4.0"
jupyter = "^1.1.1"
datasets = "^3.1.0"
accelerate = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
